mqtt reduction when lighting active, priority added

+++ code/6_Lights/03_Animator/Presets.cpp
@@ -384,7 +384,7 @@ bool mAnimatorLight::applyPreset(byte index, byte callMode)
 
 
 // apply preset or fallback to a effect and palette if it doesn't exist
-void mAnimatorLight::applyPresetWithFallback(uint8_t index, uint8_t callMode, uint8_t effectID, uint8_t paletteID)
+void mAnimatorLight::applyPresetWithFallback(uint8_t index, uint8_t callMode, uint16_t effectID, uint8_t paletteID)
 {
   applyPreset(index, callMode);  
   effectCurrent = effectID; // these two will be overwritten if preset exists in SubTask_Presets()
